Non-Unique session name for name on template
Fixed a problem where the session variables that keep entered values in
forms were the same between the 2 forms on the template page.  This led
to the arbitrary unset of the name variable for the scrape form by the
add template form.  The problem was resolved by setting the session for
name in the scrape form to something different.