URL Shortners
Added the option to shorten the phishing url using goo.gl and tiny url
to further mask links.  Google's is not working currently, but tiny url
is.  Now the URL is stored in the database and can be viewed by
hovering over the Clicked column on each row of the response page.